requirements.txt
boto3
pandas
streamlit
python-jose